get account membership from ActiveBalances
