+++ src/library/StatusLabel/index.tsx
@@ -6,11 +6,12 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { ButtonHelp } from '@polkadot-cloud/react';
 import { useHelp } from 'contexts/Help';
 import { usePlugins } from 'contexts/Plugins';
-import { usePoolMemberships } from 'contexts/Pools/PoolMemberships';
 import { useStaking } from 'contexts/Staking';
 import { useUi } from 'contexts/UI';
 import { Wrapper } from './Wrapper';
 import type { StatusLabelProps } from './types';
+import { useActiveAccounts } from 'contexts/ActiveAccounts';
+import { useBalances } from 'contexts/Balances';
 
 export const StatusLabel = ({
   title,
@@ -21,10 +22,13 @@ export const StatusLabel = ({
   status = 'sync_or_setup',
 }: StatusLabelProps) => {
   const { isSyncing } = useUi();
+  const { openHelp } = useHelp();
   const { plugins } = usePlugins();
   const { inSetup } = useStaking();
-  const { membership } = usePoolMemberships();
-  const { openHelp } = useHelp();
+  const { getPoolMembership } = useBalances();
+  const { activeAccount } = useActiveAccounts();
+
+  const membership = getPoolMembership(activeAccount);
 
   // syncing or not staking
   if (status === 'sync_or_setup') {
