+++ src/locale/cn/base.json
@@ -1,12 +1,6 @@
 {
   "base": {
     "active": "激活",
-    "allowAll": "允许所有",
-    "allowAnyoneCompound": "允许任何人代表您复利收益",
-    "allowAnyoneCompoundWithdraw": "允许任何人代表您复利或取出收益",
-    "allowAnyoneWithdraw": "允许任何人代表您取出收益",
-    "allowCompound": "允许复利",
-    "allowWithdraw": "允许取出收益",
     "community": "社区",
     "feedback": "反馈",
     "goTo": "查看",
