+++ src/locale/cn/library.json
@@ -14,6 +14,12 @@
     "address": "地址",
     "addressCopiedToClipboard": "复制到剪贴板的地址",
     "all": "全部",
+    "allowAll": "允许所有",
+    "allowAnyoneCompound": "允许任何人代表您复利收益",
+    "allowAnyoneCompoundWithdraw": "允许任何人代表您复利或取出收益",
+    "allowAnyoneWithdraw": "允许任何人代表您取出收益",
+    "allowCompound": "允许复利",
+    "allowWithdraw": "允许取出收益",
     "alreadyImported": "地址已导入",
     "asAPoolMember": "作为提名池成员",
     "asThePoolDepositor": "作为提名池存款人",
