+++ src/locale/en/base.json
@@ -1,12 +1,6 @@
 {
   "base": {
     "active": "Active",
-    "allowAll": "Allow All",
-    "allowAnyoneCompound": "Allow anyone to compound rewards on your behalf.",
-    "allowAnyoneCompoundWithdraw": "Allow anyone to compound or withdraw rewards on your behalf.",
-    "allowAnyoneWithdraw": "Allow anyone to withdraw rewards on your behalf.",
-    "allowCompound": "Allow Compound",
-    "allowWithdraw": "Allow Withdraw",
     "community": "Community",
     "feedback": "Feedback",
     "goTo": "Go To",
