+++ src/locale/en/library.json
@@ -14,6 +14,12 @@
     "address": "Address",
     "addressCopiedToClipboard": "Address Copied to Clipboard",
     "all": "All",
+    "allowAll": "Allow All",
+    "allowAnyoneCompound": "Allow anyone to compound rewards on your behalf.",
+    "allowAnyoneCompoundWithdraw": "Allow anyone to compound or withdraw rewards on your behalf.",
+    "allowAnyoneWithdraw": "Allow anyone to withdraw rewards on your behalf.",
+    "allowCompound": "Allow Compound",
+    "allowWithdraw": "Allow Withdraw",
     "alreadyImported": "Address Already Imported",
     "asAPoolMember": "as a pool member.",
     "asThePoolDepositor": "as the pool depositor.",
